requirements.txt
nicegui==2.20.0
fastapi==0.115.14
sqlalchemy==2.0.0
psycopg2-binary==2.9.10
uvicorn==0.27.0
